structure changes and other modification
