+++ imports/ui/components/admin/dashboard/dashboard.html
@@ -1,5 +1,5 @@
 <template name="Admin_dashboard">
     <div class="wrap">
-        Administrator Dashboard
+       <p>administrator</p>
     </div>
 </template>
