+++ imports/ui/components/admin/list/user-list.html
@@ -1,4 +1,4 @@
-<template name="User_list">
+<template name="Users_list">
     <div class="user-list-section page">
         <div class="content-wrap">
             <div class="content">
