--- imports/ui/components/admin/update/user-update.html
@@ -1,48 +1,0 @@
-<template name="User_update">
-    <div class="user_update_section page">
-        <div class="content-wrap">
-            <div class="content">
-                <div class="card">
-                    <form class="user-update">
-                        <div class="update_form">
-                            <div class="row">
-                                <div class="col-6">
-                                    <div class="form-group">
-                                        <label>Firstname</label>
-                                        <input type="text" class="form-control" name="firstName" value="{{update_user.profile.firstName}}">
-                                    </div>
-                                </div>
-                                <div class="col-6">
-                                    <div class="form-group">
-                                        <label>Lastname</label>
-                                        <input type="text" class="form-control" name="lastName" value="{{update_user.profile.lastName}}">
-                                    </div>
-                                </div>
-                            </div>
-                            <div class="row">
-                                <div class="col-6">
-                                    <div class="form-group">
-                                        <label>Address</label>
-                                        <input type="text" class="form-control" name="Addrewss" value="{{update_user.profile.address}}">
-                                    </div>
-                                </div>
-                                <div class="col-6">
-                                    <div class="form-group">
-                                        <label>User Type</label>
-                                        <select class="form-control" name="UserType">
-                                            <option value="admin">Admin</option>
-                                            <option value="user">User</option>
-                                        </select>
-                                    </div>
-                                </div>
-                            </div>
-                            <div class="submit-button align-right">
-                                <button class="button success">Submit</button>
-                            </div>
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </div>
-    </div>
-</template>
