+++ imports/ui/layouts/headers/admin/header.html
@@ -5,8 +5,8 @@
                 <li>
                     <a class="toggle"><i class="fa fa-bars" aria-hidden="true"></i></a>
                 </li>
-                <li class="align-right">
-                    <a href="/admin" id="logout"><i class="fa fa-share-square-o"></i>Logout</a>
+                <li class="align-right" style="padding-right: 20px;">
+                    <a href="/admin" id="logout"><i class="fa fa-share-square-o" style="font-size: 20px;padding-right: 3px;"></i>Logout</a>
                 </li>
             </ul>
         </div>
