+++ imports/ui/layouts/sidebars/admin/sidebar.html
@@ -39,5 +39,36 @@
                 </ul>
             </div>
         </div>
+        <div class="thin-menu displaynone">
+            <i class="fa fa-home"></i>
+            <div class="dropdown-content">
+                <a href="/"><span>Dashboard</span></a>
+                <div class="divide"></div>
+                <a href="/"><span>Layouts</span></a>
+            </div>
+        </div>
+        <div class="thin-menu displaynone">
+            <i class="fa fa-columns" aria-hidden="true"></i>
+            <div class="dropdown-content">
+                <a href="/"><span>Dashboard</span></a>
+                <div class="divide"></div>
+                <a href="/"><span>Layouts</span></a>
+            </div>
+        </div>
+        <div class="thin-menu displaynone">
+            <i class="fa fa-cog" aria-hidden="true"></i>
+        </div>
+        <div class="thin-menu displaynone">
+            <i class="fa fa-window-maximize" aria-hidden="true"></i>
+        </div>
+        <div class="thin-menu displaynone">
+            <i class="fa fa-commenting-o" aria-hidden="true"></i>
+        </div>
+        <div class="thin-menu displaynone">
+            <i class="fa fa-area-chart" aria-hidden="true"></i>
+        </div>
+        <div class="thin-menu displaynone">
+            <i class="fa fa-table" aria-hidden="true"></i>
+        </div>
     </div>
 </template>
