+++ imports/ui/pages/admin/home/home.js
@@ -1,1 +1,12 @@
 import './home.html';
+
+Template.Admin_home.events({
+    'click .toggle'(event, temp) {
+        temp.$('.page-sidebar').toggleClass('thin');
+        temp.$('.page-content').toggleClass('padding-left');
+        temp.$('.page-navbar').toggleClass('padding-left');
+        temp.$('.menus').toggleClass('displaynone');
+        temp.$('.navbar-logo').toggleClass('displaynone');
+        temp.$('.thin-menu').toggleClass('displayblock');
+    }
+});
