--- imports/ui/pages/admin/users/create/create.html
@@ -1,13 +1,0 @@
-<template name="User-create">
-    <main class="main">
-        <div class="page-container">
-            {{> Home_header}}
-            {{> App_sidebar}}
-            <div class="page-content">
-                <div class="wrap">
-                    {{> User_create}}
-                </div>
-            </div>
-        </div>
-    </main>
-</template>
