--- imports/ui/pages/admin/users/create/create.js
@@ -1,4 +1,0 @@
-import './create.html';
-
-// Component(s)
-import '../../../../components/admin/create/user-create.js';
