--- imports/ui/pages/admin/users/list/list.html
@@ -1,5 +1,0 @@
-<template name="Users_list">
-    <div class="wrap">
-        {{> User_list}}
-    </div>
-</template>
