--- imports/ui/pages/admin/users/list/list.js
@@ -1,4 +1,0 @@
-import './list.html';
-
-// Component(s)
-import '../../../../components/admin/list/user-list.js';
