--- imports/ui/pages/admin/users/update/update.html
@@ -1,13 +1,0 @@
-<template name="User-update">
-    <main class="main">
-            <div class="page-container">
-                {{> Home_header}}
-                {{> App_sidebar}}
-                <div class="page-content">
-                    <div class="wrap">
-                        {{> User_update}}
-                    </div>
-                </div>
-            </div>
-        </main>
-</template>
