--- imports/ui/pages/admin/users/update/update.js
@@ -1,4 +1,0 @@
-import './update.html';
-
-// Component(s)
-import '../../../../components/admin/update/user-update';
